Split GenState module into three modules.
This was necessary to avoid long compilation times for individual, which
triggered CI build timeout on hydra
