+++ libs/cardano-ledger-test/cardano-ledger-test.cabal
@@ -97,6 +97,10 @@ library
     Test.Cardano.Ledger.Rational
     Test.Cardano.Ledger.TestableEra
     Test.Cardano.Ledger.ValueFromList
+  other-modules:
+    Test.Cardano.Ledger.Generic.GenState.Types
+    Test.Cardano.Ledger.Generic.GenState.Generators
+    Test.Cardano.Ledger.Generic.GenState.GenScript
   build-depends:
     aeson >= 2,
     array,
